Microservices/monitoring/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.123.7
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
scikit-learn==1.5.2
numpy>=1.24.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
prometheus_client==0.19.0
pymongo==4.6.1
pika==1.3.2
python-socketio==5.10.0
starlette==0.50.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvicorn[standard]==0.38.0
watchfiles==1.1.1
websockets==15.0.1
